runway/requirements.txt
--extra-index-url https://nexus-tools.swacorp.com/repository/all-pypi/simple
--trusted-host nexus-tools.swacorp.com
argcomplete==1.12.2
attrs==20.3.0
awacs==1.0.2
aws-sam-translator==1.34.0
awscli==1.19.21
boto==2.49.0
boto3==1.17.21
botocore==1.20.21
ccplatlogging
certifi==2020.12.5
cffi==1.14.5
cfn-flip==1.2.3
cfn-lint==0.46.0
chardet==4.0.0
click==7.1.2
colorama==0.4.3
coloredlogs==15.0
coverage==4.3.4
cryptography==3.3.2
decorator==4.4.2
dnspython==2.1.0
docker==4.4.4
docutils==0.15.2
ecdsa==0.16.1
flake8==3.8.4
formic2==1.0.3
future==0.18.2
gitdb==4.0.5
GitPython==3.1.14
google-auth==1.27.1
humanfriendly==9.1
idna==2.10
importlib-metadata==3.7.2
iniconfig==1.1.1
Jinja2==2.11.3
jmespath==0.10.0
jq==1.1.2
jsonpatch==1.31
jsonpickle==2.0.0
jsonpointer==2.0
jsonschema==3.2.0
junit-xml==1.9
kubernetes==10.0.0
lark-parser==0.7.8
logging-formatter-anticrlf==1.2
MarkupSafe==1.1.1
mccabe==0.6.1
mock==4.0.2
more-itertools==8.7.0
moto==2.0.1
netaddr==0.8.0
networkx==2.5
nose==1.3.7
oauthlib==3.1.0
packaging==20.9
pathspec==0.8.1
pluggy==0.13.1
ply==3.11
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclean==2.0.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
pyhcl==0.4.4
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.2
pytest-cov==2.11.1
python-dateutil==2.7.5
python-hcl2==0.3.2
python-jose==3.1.0
pytz==2021.1
PyYAML==5.1.2
requests==2.25.1
requests-aws4auth==1.0.1
requests-mock==1.8.0
requests-oauthlib==1.3.0
responses==0.12.1
rsa==4.7.2
runway==1.15.0
s3transfer==0.3.4
schematics==2.0.1
semantic-version==2.8.5
semver==2.13.0
Send2Trash==1.5.0
setuptools-rust==0.11.6
six==1.15.0
smmap==3.0.5
stacker==1.7.2
stacker-blueprints==1.0.8
toml==0.10.2
troposphere==2.6.3
typing-extensions==3.7.4.3
urllib3==1.24.3
websocket-client==0.58.0
Werkzeug==1.0.1
xmltodict==0.12.0
yamllint==1.26.0
yq==2.12.0
zgitignore==1.0.0
zipp==3.4.1
